+++ src/Drivers/PrintNode/Entity/Printer.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Rawilk\Printing\Drivers\PrintNode\Entity;
 
 use Illuminate\Contracts\Support\Arrayable;