+++ src/Facades/Printing.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Rawilk\Printing\Facades;
 
 use Illuminate\Support\Facades\Facade;
@@ -12,6 +14,7 @@ use Illuminate\Support\Facades\Facade;
  * @method static \Rawilk\Printing\Contracts\PrintTask newPrintTask()
  * @method static \Rawilk\Printing\Contracts\Printer|null find($printerId = null)
  * @method static \Illuminate\Support\Collection printers()
+ * @method static \Rawilk\Printing\Printing driver(?string $driver = null)
  */
 class Printing extends Facade
 {