+++ tests/Feature/Drivers/PrintNode/PrintTaskTest.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Rawilk\Printing\Tests\Feature\Drivers\PrintNode;
 
 use Mockery;